mali: 3-step GPU DVFS control
Based on gokhanmoral's patch at:
https://github.com/gokhanmoral/siyahkernel3/commit/ad613cfea890fd52b3b7cd174bbe06463f6a1f39

With the following changes:
The second step in mali_dvfs (freq) should always be 1000000 for 1 MHz.  It is
not a voltage
Setup down threshold logic to be a target threshold - instead of being percentage
of current load, it is the desired percentage of load at the lower frequency.
This eliminates needing to think about whether your down_threshold will
be greater than 100% load at the new frequency.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>